Add more unit tests for UserService

+++ BookingWebProject.Core/Services/UserService.cs
@@ -49,7 +49,7 @@
                     City = fvh.Hotel.City,
                     Country = fvh.Hotel.Country,
                     IsFavorite = true,
-                    PicturePath = fvh.Hotel.Pictures.First().Path,
+                    PicturePath = fvh.Hotel.Pictures.FirstOrDefault().Path,
                     CheapestHotelRoomViewModel = fvh.Hotel.Rooms
                    .Where(r => !r.IsDeleted && !r.RoomType.IsDeleted)
                    .Select(r => new CheapestHotelRoomViewModel() { Id = r.Id, PricePerNight = r.PricePerNight })
@@ -91,17 +91,20 @@
 
         public async Task DeleteUserProfilePictureAsync(Guid userId, string path)
         {
-            string profilePictureName = path.Split("\\")[2];
-            //Set your path when use the application;
-            string profilePicturesFolderPath = Path.GetFullPath(@"C:\Users\Home\Desktop\Booking Web Project C# Web\BookingWebProject\BookingWebProject\wwwroot\img\ProfilePictures\");
-            string[] files = Directory.GetFiles(profilePicturesFolderPath);
-            if (files.Length > 0)
+            if (!string.IsNullOrWhiteSpace(path))
             {
-
-                string fileToDelete = files.FirstOrDefault(f => f.EndsWith(profilePictureName));
-                if (fileToDelete != null)
+                string profilePictureName = path.Split("\\")[2];
+                //Set your path when use the application;
+                string profilePicturesFolderPath = Path.GetFullPath(@"C:\Users\Home\Desktop\Booking Web Project C# Web\BookingWebProject\BookingWebProject\wwwroot\img\ProfilePictures\");
+                string[] files = Directory.GetFiles(profilePicturesFolderPath);
+                if (files.Length > 0)
                 {
-                    File.Delete(fileToDelete);
+
+                    string fileToDelete = files.FirstOrDefault(f => f.EndsWith(profilePictureName));
+                    if (fileToDelete != null)
+                    {
+                        File.Delete(fileToDelete);
+                    }
                 }
             }
             User user = await bookingContext.Users
